feat(discoveryProxyDataLoader): moved reused function processAdvancedSearchTerms to utils file
